Added v4 API Key support with tests
